Add trash alias to bashrc

+++ bashrc
@@ -14,8 +14,13 @@ wine_pgrep() {
 export PS1='\n\[\e]0;\u@\h: \w\a\]${debian_chroot:+($debian_chroot)}\[\033[01;32m\]\u@\h\[\033[00m\]:\[\033[01;34m\]\w\[\033[00m\]\n\$'
 export EDITOR=vim
 
+if command -v 'trash' &> /dev/null; then
+    alias rm='trash'
+else
+    alias rm='rm -i'
+fi
+
 alias cp='cp -i'
-alias rm='rm -i'
 alias mv='mv -i'
 alias screen='tmux attach -t'
 alias hex="printf '0x%x\\n'"
